+++ tests/rename.annotated.js
@@ -337,6 +337,38 @@ obj = /*@ngInject*/ {
     but: { onlythrough: ["object literals", {donttouch: function(me) {}}]},
 };
 
+obj = {
+    /*@ngInject*/
+    foo: ["a", function(a) {}],
+    bar: function(b, c) {},
+};
+
+/*@ngInject*/
+obj = {
+    foo: ["a", function(a) {}],
+    bar: ["b", "c", function(b, c) {}],
+    val: 42,
+    inner: {
+        circle: ["d", function(d) {}],
+        alalalala: "long",
+    },
+    nest: { many: {levels: ["x", function(x) {}]}},
+    but: { onlythrough: ["object literals", {donttouch: function(me) {}}]},
+};
+
+/*@ngInject*/
+var obj = {
+    foo: ["a", function(a) {}],
+    bar: ["b", "c", function(b, c) {}],
+    val: 42,
+    inner: {
+        circle: ["d", function(d) {}],
+        alalalala: "long",
+    },
+    nest: { many: {levels: ["x", function(x) {}]}},
+    but: { onlythrough: ["object literals", {donttouch: function(me) {}}]},
+};
+
 // @ngInject
 function foo($scope) {
 }
@@ -350,72 +382,89 @@ Foo.$inject = ["$scope"];
 
 // @ngInject
 // has trailing semicolon
-var foo = function($scope) {
+var foo1 = function($scope) {
 };
-foo.$inject = ["$scope"];
+foo1.$inject = ["$scope"];
 
 // @ngInject
 // lacks trailing semicolon
-var foo = function($scope) {
+var foo2 = function($scope) {
 }
-foo.$inject = ["$scope"];
+foo2.$inject = ["$scope"];
 
 // @ngInject
 // has trailing semicolon
-bar.foo = function($scope) {
+bar.foo1 = function($scope) {
 };
-bar.foo.$inject = ["$scope"];
+bar.foo1.$inject = ["$scope"];
 
 // @ngInject
 // lacks trailing semicolon
-bar.foo = function($scope) {
+bar.foo2 = function($scope) {
 }
-bar.foo.$inject = ["$scope"];
+bar.foo2.$inject = ["$scope"];
 
 // let's zip-zag indentation to make sure that the $inject array lines up properly
     // @ngInject
-    function foo($scope) {}
-    foo.$inject = ["$scope"];
+    function foo1($scope) {}
+    foo1.$inject = ["$scope"];
         // @ngInject
-        function foo($scope) {
+        function foo2($scope) {
         }
-        foo.$inject = ["$scope"];
-/* @ngInject */ function foo($scope) {}
-foo.$inject = ["$scope"];
-            /* @ngInject */ function foo($scope) {
+        foo2.$inject = ["$scope"];
+/* @ngInject */ function foo3($scope) {}
+foo3.$inject = ["$scope"];
+            /* @ngInject */ function foo4($scope) {
             }
-            foo.$inject = ["$scope"];
+            foo4.$inject = ["$scope"];
 
     // @ngInject
-    var foo = function($scope) {
+    var foo1 = function($scope) {
     };
-    foo.$inject = ["$scope"];
+    foo1.$inject = ["$scope"];
         // @ngInject
-        var foo = function($scope) {};
-        foo.$inject = ["$scope"];
+        var foo2 = function($scope) {};
+        foo2.$inject = ["$scope"];
 // @ngInject
-var foo = function($scope) {
+var foo3 = function($scope) {
 }
-foo.$inject = ["$scope"];
+foo3.$inject = ["$scope"];
             // @ngInject
-            var foo = function($scope) {}
-            foo.$inject = ["$scope"];
+            var foo4 = function($scope) {}
+            foo4.$inject = ["$scope"];
 
-    /* @ngInject */ var foo = function($scope) {
+    /* @ngInject */ var foo5 = function($scope) {
     };
-    foo.$inject = ["$scope"];
-        /* @ngInject */var foo = function($scope) {};
-        foo.$inject = ["$scope"];
-/* @ngInject */var foo = function($scope) {
+    foo5.$inject = ["$scope"];
+        /* @ngInject */var foo6 = function($scope) {};
+        foo6.$inject = ["$scope"];
+/* @ngInject */var foo7 = function($scope) {
 }
-foo.$inject = ["$scope"];
-            /* @ngInject */var foo = function($scope) {}
-            foo.$inject = ["$scope"];
+foo7.$inject = ["$scope"];
+            /* @ngInject */var foo8 = function($scope) {}
+            foo8.$inject = ["$scope"];
 
 
 // adding an explicit annotation where it isn't needed should work fine
 myMod.controller("foo", /*@ngInject*/ ["$scope", "$timeout", function($scope, $timeout) {
 }]);
+
+
+
+// explicit annotations using ngInject() instead of /*@ngInject*/
+var x = ngInject(["$scope", function($scope) {}]);
+
+obj = ngInject({
+    foo: ["a", function(a) {}],
+    bar: ["b", "c", function(b, c) {}],
+    val: 42,
+    inner: {
+        circle: ["d", function(d) {}],
+        alalalala: "long",
+    },
+    nest: { many: {levels: ["x", function(x) {}]}},
+    but: { onlythrough: ["object literals", {donttouch: function(me) {}}]},
+});
 
 
 
@@ -446,3 +495,20 @@ var x = /*@ngInject*/ (function() {
     return ["$aRenamed", function($a) {
     }];
 })();
+
+
+// reference support
+function MyCtrl1(a, b) {
+}
+MyCtrl1.$inject = ["a", "b"];
+if (true) {
+    // proper scope analysis including shadowing
+    let MyCtrl1 = ["c", function(c) {
+    }]
+    angular.module("MyMod").directive("foo", MyCtrl1);
+}
+angular.module("MyMod").controller("bar", MyCtrl1);
+function MyCtrl2(z) {
+}
+MyCtrl2.$inject = ["z"];
+funcall(/*@ngInject*/ MyCtrl2); // explicit annotation on reference flows back to definition